[refact] String Extension
body: utf8 로 인코딩 디코딩 하는 Extension

Related to: #30
